Add support for unknown IPL3 variants
Currently, CIC autodetection features works by calculating a CRC32 of
the IPL3 boot code (header bytes 0x40..0x1000) and checking it against
a known database to know which CIC is expected. This approach cannot
detect unknown IPL3s and is going to be insufficient in the short
future when libdragon releases its own open source IPL3.

A more robust approach is to use the same custom checksum function
used by IPL2, trying to simulate boot with different CICs: that is,
try with different seeds and see if the respective checksum matches.
This basically means that we now have a DB of CIC secrets rather than
IPL3 CRCs: it seems a better compromise if several new IPL3 spawn
from libdragon's.